+++ classes/patient.py
@@ -129,9 +129,9 @@ class Patient():
             scenario.lsoa_mt_travel_time[self.lsoa])
 
         self.mt_transfer_unit = (
-            scenario.national_mt_transfer_unit[self.closest_ivt_unit])
+            scenario.national_dict['mt_transfer_unit'][self.closest_ivt_unit])
         self.mt_transfer_travel_duration = (
-            scenario.national_mt_transfer_time[self.closest_ivt_unit])
+            scenario.national_dict['mt_transfer_time'][self.closest_ivt_unit])
         self.mt_transfer_required = (
             self.closest_ivt_unit != self.closest_mt_unit)
